+++ src/Redux/categories/categories.js
@@ -2,7 +2,7 @@ const CHECK_STATUS = 'bookstore/categories/CHECK_STATUS';
 
 const categoriesReducer = (state = [], action) => {
   if (action.type === CHECK_STATUS) {
-    return ['Under Construction'];
+    return 'UNDER CONSTRUCTION';
   }
   return state;
 };
