+++ src/Redux/configureStore.js
@@ -1,12 +1,19 @@
-import { configureStore } from '@reduxjs/toolkit';
-import bookReducer from './books/books';
-import categoriesReducer from './categories/categories';
+/*eslint-disable*/
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import './index.css';
+import App from './App';
+import store from './redux/configureStore';
 
-const store = configureStore({
-  reducer: {
-    book: bookReducer,
-    category: categoriesReducer,
-  },
-});
-
-export default store;
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </Provider>
+  </React.StrictMode>,
+);
